+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## Upcoming
 
-## [0.7.2](https://github.com/whitefusionhq/esbuild-plugin-html-modules/compare/v0.7.0...v0.7.1) - 2024-13-06
+## [0.8.0](https://github.com/whitefusionhq/esbuild-plugin-html-modules/compare/v0.7.2...v0.8.0) - 2024-10-27
+
+- Remove the `extractScopedStyles: true` option and dependency on `enhance-style-transform`.
+- Add support for `<style global>` syntax in `extractGlobalStyles`
+
+## [0.7.2](https://github.com/whitefusionhq/esbuild-plugin-html-modules/compare/v0.7.0...v0.7.1) - 2024-03-06
 
 - Fix bug where the `ignoreSSROnly` option wasn't including client-side scripts when the HTML was totally blank